pyproject.toml
[tool.poetry]
name = "fabriquetest-repo"
version = "0.1.0"
description = ""
authors = ["quanttyo <divergere.ratio@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "2.2.10"
django-rest-framework = "^0.1.0"
django-rest-swagger = "^2.2.0"
django-queryset-exts = "^0.0.1"
drf-optional-fields = "^0.0.1"
flake8 = "^3.9.2"
black = "^21.6b0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.1"
django-environ = "^0.4.5"
djangorestframework-simplejwt = "^4.7.1"
jsonfield = "^3.1.0"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
